MSVC had twenty more, behind the first one
/WX stops the compiler at the first warning, so every CI run since the checkout
started working had reported exactly one name and hidden the rest. Visual Studio
2022 is installed on this machine; building with it locally turned a five-minute
round trip per warning into a rebuild.

Conversions MSVC narrows and GCC does not: u32 owned_bit into a u16 weapons
mask, s32 ammo_type into an s8, a u32 qscale into the frame's u16. Each is
deliberate and each now says so.

C4090, three times in the tools and twice in creworld: `memset` on an array of
pointers-to-const. MSVC reads reaching a `void *` parameter as discarding a
qualifier. The arrays are locals and zeroing them is exactly what is meant.

C4701 seven more times, always the same shape this round has met on every
compiler -- a flag guards a value and the compiler does not correlate the two.
Zeroed at declaration, except world.c's sort_screen_state, where the record
array is far too big to zero every call for a warning and only the cursor is
read under the flag.

And two are turned off, each for a stated reason rather than to make the build
quiet. C4100 is the unreferenced parameter this project already ignores on GCC
and Clang, and it should not depend on which compiler is running. C4127 fires on
every CHECK(Q2_MISSION_BODY_Y == 36, ...) in the test suites -- fifteen of them
before the build even reached the tools -- and asserting the constants read out
of the executable is what those suites exist to do. Where C4310 caught a real
narrowing in surface.h it stays on; the two tests that deliberately demonstrate
a wrap now cast a variable rather than a literal.

Four compilers clean with -Werror: GCC 13.3, Clang on Linux and Windows, MSVC
19.44. 35 C tests on each, 54 tooling tests.

+++ src/game/simcombat.c
@@ -249,7 +249,8 @@ void q2_sim_combat_init(q2_sim *sim)
     /* A fresh player has the blaster and nothing else, which is what the
      * spawn path at 0x8003D4FC leaves in the weapon fields. */
     sim->combat.weapon_id        = Q2_WID_BLASTER;
-    sim->combat.inv.weapons      = q2_weapon_tables_builtin()->owned_bit[Q2_WID_BLASTER];
+    sim->combat.inv.weapons      =
+        (u16)q2_weapon_tables_builtin()->owned_bit[Q2_WID_BLASTER];
     sim->combat.chaingun_bullets = 1;
 
     q2_actor_from_player(&sim->combat.self, &sim->combat.inv, sim->player[sim->cur_player].pos);
@@ -464,7 +465,7 @@ q2_fire_result_v2 q2_sim_fire(q2_sim *sim)
      * see weapon.h. Drawn AFTER the shot so the flash does not perturb the
      * spread sequence, which is what the fire function's own ordering does.
      */
-    if (q2_weapon_has_muzzle_light(sim->combat.weapon_id)) {
+    if (q2_weapon_has_muzzle_light((u8)sim->combat.weapon_id)) {
         static const u8 flash[3] = { Q2_MUZZLE_LIGHT_R, Q2_MUZZLE_LIGHT_G,
                                      Q2_MUZZLE_LIGHT_B };
         s32 inner, outer;
